Fixed Level.java issue

+++ src/java/org/blockserver/blocks/Block.java
@@ -1,8 +1,8 @@
 package org.blockserver.blocks;
 
 public class Block{
-	private int id;
-	private int damage;
+	private byte id;
+	private byte damage;
 	private String name;
 
 	public Block(int id){
@@ -15,19 +15,18 @@ public class Block{
 		this(name, id, 0);
 	}
 	public Block(String name, int id, int damage){
-		this.id = id;
-		this.damage = damage;
+		this.id = (byte) id;
+		this.damage = (byte) damage;
 		this.name = name.isEmpty() ? "Unknown":name;
 	}
 
-	public int getID(){
+	public byte getID(){
 		return id;
 	}
-	public int getDamage(){
+	public byte getDamage(){
 		return damage;
 	}
 	public String getName(){
 		return name;
 	}
-
 }
